backend/requirements.txt
# AI Debate Platform - Backend Dependencies

# Core Framework
Flask==3.0.0
flask-cors==4.0.0

# AI/ML - LATEST VERSION (google-genai replaces google-generativeai)
google-genai>=0.3.0
scikit-learn==1.3.2
numpy==1.24.3

# Database
pymongo==4.6.1

# Authentication
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0

# Environment
python-dotenv==1.0.0

# Optional: For better performance
gunicorn==21.2.0
